log: Mitigate "exactly full" corner case
I've seen log draining lock up, which could be because if the buffer
ends up exactly full, extract_idx == insert_idx.

Tweak the make space algorithm to ensure there's always at least 1 byte
free - which is wasteful, but avoids this issue.

An improvement would be to switch the indices to not wrap on the buffer
size, and only modulo them for buffer access.