WIP: Use PIO and DMA for frame capture
I'd like to find a way to unpack the YUV into planes. To do it via
DMA I think it would have to be via a DMA control block *per byte*
which is pretty awful.

It might be possible to DMA word-wise to a second state machine which
can then shuffle the data.