WIP: Hack together shift-register-based camera
Uses 13 instructions which feels a bit high, but likely fine.

We are now using X for counting bits, which means there's only one
scratch register available for row/col loops - therefore we have to send
the number of columns on every row, meaning more host involvement.
Rows are ~3ms apart, so this shouldn't be the end of the world.

Shift register wiring:
- CLK is sourced from XCLK, accessed in PIO via internal GPIO routing
- SH/nLD is connected to PXCLK
- CLK_INH is wired to an ouput GPIO (currently the old D1)
- QH is wired to an input GPIO (currently D0)

Then the Pico needs:
- SDA/SCL
- XCLK
- VSYNC
- HREF
- PXCLK
- CLK_INH
- QH

Saving of 6 pins compared to parallel input.