camera: Somewhat reliable on-board.
This was quite a saga. Wrapping the ribbon cable in foil helps a lot
with some noise issue. I may have also had some issue with crosstalk on
the wires on stripboard.

I've also dropped the clock for better reliability, and adjusted timing
a bit.

With shorter/better cables maybe we can go full speed again, but I want
to take a look on a high speed scope to tune timing.