heading: Add some nasty special casing to improve turning
It's been a constant problem that the controller gets stuck with a small
turn angle. This adds some special casing to "nudge" the last little
bit.

It seems quite effective. I thought this was kinda what the integral
controller term was meant to do, but I haven't managed to find a tuning
that works.